Module Input.Tbl

Table to associate value to positions in input buffers. The complexity of access in the table is O(ln(N)) where N is the number of tables.

type 'a t

val create : unit -> 'a t

val add : 'a t -> buffer -> idx -> 'a -> unit

val find : 'a t -> buffer -> idx -> 'a

val clear : 'a t -> unit

val iter : 'a t -> ('a -> unit) -> unit