Parameter Make.S

val id_charset : Charset.t

id_charset contains the characters that are not allowed to directly follow a keyword (because it would make it an identifier for example). In general, it should correspond to characters that are can be used in identifiers.

val reserved : string list

reserved is a liste of words that must be rejected as identifiers as if they were keywords. Note that if a word is in the list, it will not be possible to define it as a keyword.